add stats handler to compute stats
